Module trust_dns_openssl::tls_server

Structs

PKey
A public or private key.

ParsedPkcs12

Pkcs12

Stack
An owned stack of T.

X509
An X509 public key certificate.

Enums

Private
A tag type indicating that a key has private components.

Traits

SslAcceptorExt
Extension trait for the SslAcceptor type in the openssl crate.

Functions

new_acceptor
Construct the new Acceptor with the associated pkcs12 data

read_cert_pem
Read the certificate from the specified path.

read_cert_pkcs12
Read the certificate from the specified path.

read_key_from_der
Reads a private key from a der formatted file

read_key_from_pkcs8
Reads a private key from a pkcs8 formatted, and possibly encoded file